+++ client/src/components/App/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import './App.css'
 import Register from '../Register/Register'
-import { Routes, Route } from 'react-router-dom'
+import { Routes, Route, Router } from 'react-router-dom'
 import Home from '../Home/Home'
 import Store from '../Store/Store'
 import Details from '../Details/Details'
@@ -11,13 +11,14 @@ import Login from '../Login/Login'
 import PasarelaStripe from '../Pasarela/PasarelaStripe'
 import FinalRegister from '../FinalRegister/FinalRegister'
 import Cart from '../Cart/Cart'
-
+import Registro from '../GoogleRegister/registro'
 
 
 function App() {
   return (
     <div>
       <Routes>
+        <Route path='/test' element={<Registro/>}></Route>
         <Route path='/pasarela' element={<PasarelaStripe/>}/>
         <Route path='/' element={<Home />} />
         <Route path='/store' element={<Store />} />
